feat(browser): use browser-compatible dashsight, (re)move non-browser cruft
