+++ bin/_qr.js
