+++ lib/browser-request.js
@@ -1,0 +1,4 @@
+"use strict";
+
+// provide a standards-compliant user-agent
+module.exports = require("@root/request");
