+++ lib/crowdnode.js
@@ -8,8 +8,8 @@ const DUFFS = 100000000;
 
 let Dash = require("./dash.js");
 let Dashcore = require("@dashevo/dashcore-lib");
-let Insight = require("./insight.js");
-let Ws = require("./ws.js");
+let Insight = require("dashsight");
+let Ws = require("dashsight/ws");
 
 CrowdNode._insightBaseUrl = "";
 // TODO don't require these shims
