+++ lib/request.js
@@ -1,5 +1,6 @@
 "use strict";
 
+//@ts-ignore
 let pkg = require("../package.json");
 
 // provide a standards-compliant user-agent
